Rename _setvars.bat to more standard activate.bat
In fact, what we are doing here is very much similar to a homebrewn
virtualenv, and we might be better off just using virtualenv/venv at
this point.

Note the original reason not to use virtualenv was that it is basically
not relocatable. However, we don't need portable/relocatability anymore
now that installation methods have improved by the widespread
availability of wheel installers.