Use j2cli to avoid open bug in jinja2-cli
jinja2-cli is currently unusable on windows due to a bug that raises an
exception whenever the utility is called without an env file. While the
bug the bug has been fixed in the source code for more than two years
thanks to a user contribution, the maintainer apparently doesn't have
the time to make a new release.

This should be reverted once the package is better maintained (as I
prefer explicitly passing parameters via -DVAR=value over implicitly
providing the whole environment as context).